requirements.txt
# Airflow requirements for Google Cloud Platform
apache-airflow-providers-google

# Reading parquet files
pyarrow